// src/utils/enum.ts
export enum ErrorMessage {
  FAILED_TO_POST_BLOG = "Failed To Create Blog ",
  FAILED_TO_CONNECT_DB = "Failed to Connect DataBase",
  EMAIL_ALREADY_EXIST = "Email already Exist",
  FAILED_TO_CREATE_USER = "Failed To create User",
  FAILED_TO_LOGIN = "Failed To Login",
  INCORRECT_PASSWORD = "Incorrect Password",
  USER_NOT_FOUND = "User Not Found",
  TOKEN_REQUIRED = "Token Required",
  SESSION_EXPIRED = "Session Expired",
  INVALID_BLOGID = "Invalid BlogId",
  FAILED_TO_DELETE_BLOG = "Failed To Delete Blog",
  FAILED_TO_UPDATE_BLOG = "Failed To update Blog",
}

export enum SuccessMessage {
  BLOG_CREATED_SUCCESSFULLY = "Blog Created Successfully",
  DB_CONNECTED_SUCCESSFULLY = "DataBase Connected Successfully",
  USER_CREATED_SUCCESSFULLY = "User Created Successfully",
  LOGIN_SUCCESSFULL = "User Logged In successfully",
  BLOG_DELETED_SUCCESSFULLY = "Blog Deleted Successfully",
  BLOG_UPDATED_SUCCESSFULLY = "Blog Updated Successfully",
}